+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Go: Update dependencies.
+- Client: Some improvements. ([#619](https://github.com/giantswarm/clustertest/pull/619))
+  - Conditions: Rework `IsClusterReadyCondition`.
+  - Client: Rework `NewFromSecret`.
+  - Client: Rework `GetClusterKubeConfig`.
+  - Client: Rework `getTeleportKubeConfig`.
+  - Client: Rework `getCAPIKubeConfig`.
 
 ## [3.1.2] - 2026-02-03